requirements.txt
streamlit
pillow
python-dotenv
groq
paddleocr
paddlepaddle==2.6.1
opencv-python-headless
numpy
graphviz